refactor: moved code into committee.go
